+++ domain/src/main/kotlin/de/ma/tracker/domain/state/message/StateCreate.kt
@@ -3,6 +3,6 @@ package de.ma.tracker.domain.state.message
 import java.time.LocalDateTime
 
 interface StateCreate {
-    val price: Long
+    val price: Float
     val entryDate: LocalDateTime?
 }
